Yield Upgrade and Automation System Implemented
Infinite yield upgrade shop is implemented to spend gold to get more yields per harvest. And automation system where you can unlock it once and toggle it on and off. It remembers the last planted seed, plants and harvest it repeatedly. If you are out of money or seeds selling crops and buying seeds for you.